requirements.txt
anyio==3.6.2
async-timeout==4.0.2
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.89.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-datastore==2.12.0
google-cloud-ndb==2.1.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httptools==0.5.0
idna==3.4
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pymemcache==4.0.0
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
